Add remaining properties to User

+++ src/main/java/org/mascotadopta/adoptionsplatform/adoptions/AdoptionsRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
  * Adoptions' data retrieval functionality.
  */
 @Repository
-public interface AdoptionsRepository extends CrudRepository<AdoptionApplication, Integer>
+public interface AdoptionsRepository extends CrudRepository<AdoptionApplication, Long>
 {
 }
