+++ src/main/java/org/mascotadopta/adoptionsplatform/pets/PetsRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
  * Pets' data retrieval functionality.
  */
 @Repository
-public interface PetsRepository extends CrudRepository<Pet, Integer>
+public interface PetsRepository extends CrudRepository<Pet, Long>
 {
 }
